Revert "Plugins: Forbid loading Angular plugins when Angular is disabled (#69225)"
This reverts commit ff34279ff43a8660e8445c8f320390ed61bcd0e8.
