+++ pkg/plugins/manager/loader/angulardetector/angulardetector_test.go
@@ -44,10 +44,9 @@ func TestAngularDetector_Inspect(t *testing.T) {
 		},
 		exp: false,
 	})
-	inspector := NewDefaultPatternsListInspector()
 	for _, tc := range tcs {
 		t.Run(tc.name, func(t *testing.T) {
-			isAngular, err := inspector.Inspect(tc.plugin)
+			isAngular, err := Inspect(tc.plugin)
 			require.NoError(t, err)
 			require.Equal(t, tc.exp, isAngular)
 		})
@@ -55,33 +54,7 @@ func TestAngularDetector_Inspect(t *testing.T) {
 
 	t.Run("no module.js", func(t *testing.T) {
 		p := &plugins.Plugin{FS: plugins.NewInMemoryFS(map[string][]byte{})}
-		_, err := inspector.Inspect(p)
+		_, err := Inspect(p)
 		require.ErrorIs(t, err, plugins.ErrFileNotExist)
-	})
-}
-
-func TestFakeInspector(t *testing.T) {
-	t.Run("FakeInspector", func(t *testing.T) {
-		var called bool
-		inspector := FakeInspector{InspectFunc: func(p *plugins.Plugin) (bool, error) {
-			called = true
-			return false, nil
-		}}
-		r, err := inspector.Inspect(&plugins.Plugin{})
-		require.True(t, called)
-		require.NoError(t, err)
-		require.False(t, r)
-	})
-
-	t.Run("AlwaysAngularFakeInspector", func(t *testing.T) {
-		r, err := AlwaysAngularFakeInspector.Inspect(&plugins.Plugin{})
-		require.NoError(t, err)
-		require.True(t, r)
-	})
-
-	t.Run("NeverAngularFakeInspector", func(t *testing.T) {
-		r, err := NeverAngularFakeInspector.Inspect(&plugins.Plugin{})
-		require.NoError(t, err)
-		require.False(t, r)
 	})
 }
