+++ pkg/services/pluginsintegration/config/config.go
@@ -41,7 +41,6 @@ func ProvideConfig(settingProvider setting.Provider, grafanaCfg *setting.Cfg, fe
 		grafanaCfg.PluginsCDNURLTemplate,
 		tracingCfg,
 		featuremgmt.ProvideToggles(features),
-		grafanaCfg.AngularSupportEnabled,
 	), nil
 }
